все даблы заменил на флоты. Добавил цвет текстур, вылез баг с отрисовкой стен

+++ src/aux.c
@@ -1,6 +1,6 @@
 #include "wolf3d.h"
 
-double arcToRad(double angle)
+float arcToRad(float angle)
 {
 	return ((angle * M_PI) / a.a180);
 }
@@ -10,7 +10,7 @@ int		color_to_hex(int r, int g, int b)
 	return (r << 16) | (g << 8) | b;
 }
 
-int	add_arc(double *arc, double to_add)
+int	add_arc(float *arc, float to_add)
 {
 	*arc += to_add;
